+++ CHANGELOG.md
@@ -24,10 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - expnses schema in `src/api/v1/apps/expenses/schemas.py`
 
 ### Changed
-- N/A
+- Added expenses field to the template in `src/templates/expenses/expenses.html`
 
 ### Fixed
-- N/A
+- Added supplier field at all expenses test
 
 ## [3.1.0] 2026-08-30
 This tag add all the serialization to save and retrieve photos of a product